Split form build methods into their own build class
